MeetVideo에서 사용되는 인터페이스 types 폴더의 meet.ts 로 분리
