+++ client/src/components/Meet/MeetVideo/FocusedVideo/index.tsx
@@ -1,9 +1,10 @@
 import React, { useEffect, useRef } from 'react';
 
 import { MicOffIcon, SpeakerOffIcon } from '@components/common/Icons';
-import { SelectedVideo } from '..';
+
 import { ThumbnailWrapper, Thumbnail } from '../style';
 import { UserInfo, Video, VideoWrapper, DeviceStatus } from './style';
+import { SelectedVideo } from '@customTypes/meet';
 
 function FocusedVideo({
   selectedVideo: { loginID, username, stream, thumbnail, mic, cam, speaker, isScreen },
