+++ client/src/components/Meet/MeetVideo/OtherVideo/index.tsx
@@ -1,7 +1,7 @@
 import React, { useRef, useEffect } from 'react';
 
 import { CameraOnIcon, MicOffIcon, SpeakerOffIcon } from '@components/common/Icons';
-import { MeetingMember, SelectedVideo } from '..';
+
 import {
   VideoWrapper,
   Video,
@@ -10,6 +10,7 @@ import {
   ThumbnailWrapper,
   SelectVideoIndicator,
 } from '../style';
+import { MeetingMember, SelectedVideo } from '@customTypes/meet';
 
 function OtherVideo({
   member: { socketID, loginID, username, thumbnail, cam, speaker, mic, stream, screen, pc },
